requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.4; python_version >= '3.6'
dj-rest-auth==2.1.5
django-cors-headers==3.7.0
django-environ==0.4.5
django==3.2.3
djangorestframework-simplejwt==4.7.0
djangorestframework==3.12.4
mysqlclient==2.0.3
pillow==8.2.0
pyjwt==2.1.0; python_version >= '3.6'
pytz==2021.1
sqlparse==0.4.1; python_version >= '3.5'
